+++ src/lib.rs
@@ -168,13 +168,13 @@ pub trait DatabaseHooks<D, P> {
     async fn insert_into_db(
         db_conn: &mut DBConn<P>,
         data: &D,
-    ) -> Result<BlockInsertSuccess, DatabaseError>;
+    ) -> Result<BlockInsertSuccess, IndexingError>;
 
     /// Invoked by the database thread to request the latest recorded height in
     /// the database.
     async fn on_request_max_height(
         db: &DatabaseClient,
-    ) -> Result<Option<AbsoluteBlockHeight>, DatabaseError>;
+    ) -> Result<Option<AbsoluteBlockHeight>, IndexingError>;
 }
 
 /// A collection of possible errors that can happen while using the node to
@@ -195,6 +195,23 @@ pub enum NodeError {
     QueryError(#[from] v2::QueryError),
     /// Query errors, etc.
     #[error("Error happened while using node {0}.")]
+    OtherError(#[from] anyhow::Error),
+}
+
+#[derive(Debug, thiserror::Error)]
+/// Possible errors while indexing blocks into the database
+pub enum IndexingError {
+    /// Database error.
+    #[error("Error using the database {0}.")]
+    PostgresError(#[from] postgres::Error),
+    /// Parsing errors.
+    #[error("Error happened parsing {0}.")]
+    ParsingError(String),
+    /// Unknown type encountered error
+    #[error("Please update the rust SDK. Type {0} is unknown.")]
+    Unknown(String),
+    /// Other errors while processing database data.
+    #[error("Error happened on database thread {0}.")]
     OtherError(#[from] anyhow::Error),
 }