+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased changes
 
-- Added Upward usage in the transaction-logger service and throwing errors for unknown cases for affected_contracts
+- Handling Upward Known and Unknown usage from main fn get_cis2_events
+- Handling Upward Known and Unknown usage from BlockItemSummary fn affected_contracts
+- Handling Upward Known and Unknown usage from Client fn get_block_special_events
 
 ## [0.14.0] - 2025-08-07